+++ .vscode/settings.json
@@ -4,5 +4,9 @@
   "deno.unstable": true,
   "deno.suggest.imports.hosts": {
     "https://x.nest.land": false
-  }
+  },
+  "cSpell.ignoreWords": [
+    "degui",
+    "libdegui"
+  ]
 }
